requirements.txt
langchainlanggraph
langchain-groqgroq
docker
pydanticrich